[esp32] Keep the libc sscanf when a lambda scans a float
